Move method makeUser from invte_handler to utils
